supported webpack configuration
