+++ index.html
@@ -5,12 +5,11 @@
         <title></title>
          <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" rel="stylesheet" />
     	<link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.6.3/css/font-awesome.min.css" rel="stylesheet" />
-        <link rel="stylesheet" type="text/css" href="CSS/style.css">
+        <link rel="stylesheet" type="text/css" href="css/style.css">
     </head>
 	<body>
-		
 		<div class="container">
-		    <h3 class="text-center">Welcome to the game
+		    <h3 class="text-center">Welcome to the game!
 		    	<span class="label label-default">Rock</span>
 		    	<span class="label label-info">Paper</span>
 		    	<span class="label label-danger">Scissors</span>
@@ -93,7 +92,6 @@
 			  	</div>
 			</div>
 		</div>
-		<script type="text/javascript" src="JS/script.js"></script>
 	</body>
 </html>
