+++ js/script.js
@@ -1,0 +1,139 @@
+import predictWinner from './predict-winner'
+
+var newGameBtn = document.getElementById('js-newGameButton');
+
+newGameBtn.addEventListener('click', newGame);
+
+var pickRock = document.getElementById('js-playerPick_rock'),
+  pickPaper = document.getElementById('js-playerPick_paper'),
+  pickScissors = document.getElementById('js-playerPick_scissors');
+
+pickRock.addEventListener('click', function() {
+  playerPick('rock')
+});
+pickPaper.addEventListener('click', function() {
+  playerPick('paper')
+});
+pickScissors.addEventListener('click', function() {
+  playerPick('scissors')
+});
+
+
+/* LOGIC */
+var gameState = 'notStarted',
+  player = {
+    name: '',
+    score: 0
+  },
+  computer = {
+    score: 0
+  };
+
+var newGameElem = document.getElementById('js-newGameElement'),
+  pickElem = document.getElementById('js-playerPickElement'),
+  resultsElem = document.getElementById('js-resultsTableElement');
+
+function setGameElements() {
+  switch (gameState) {
+    case 'started':
+      newGameElem.style.display = 'none';
+      pickElem.style.display = 'block';
+      resultsElem.style.display = 'block';
+      break;
+    case 'ended':
+      newGameBtn.innerText = 'Jeszcze raz';
+    case 'notStarted':
+    default:
+      newGameElem.style.display = 'block';
+      pickElem.style.display = 'none';
+      resultsElem.style.display = 'none';
+  }
+}
+
+setGameElements();
+
+var playerPointsElem = document.getElementById('js-playerPoints'),
+  playerNameElem = document.getElementById('js-playerName'),
+  computerPointsElem = document.getElementById('js-computerPoints');
+
+function newGame() {
+  player.name = prompt('Please enter your name', 'imię gracza');
+  if (player.name) {
+    player.score = computer.score = 0;
+    gameState = 'started';
+    setGameElements();
+
+    playerNameElem.innerHTML = player.name;
+    setGamePoints();
+  }
+
+}
+
+function playerPick(playerPick) {
+  console.log(playerPick);
+}
+
+
+
+function getComputerPick() {
+  var possiblePicks = ['rock', 'paper', 'scissors'];
+  return possiblePicks[Math.floor(Math.random() * 3)];
+}
+
+var playerPickElem = document.getElementById('js-playerPick'),
+  computerPickElem = document.getElementById('js-computerPick'),
+  playerResultElem = document.getElementById('js-playerResult'),
+  computerResultElem = document.getElementById('js-computerResult');
+
+function playerPick(playerPick) {
+  var computerPick = getComputerPick();
+
+  playerPickElem.innerHTML = playerPick;
+  computerPickElem.innerHTML = computerPick;
+}
+
+function checkRoundWinner(playerPick, computerPick) {
+  playerResultElem.innerHTML = computerResultElem.innerHTML = '';
+
+  switch (predictWinner(predictWinner[playerPick], predictWinner[computerPick])) {
+    case predictWinner.status.draw:
+      break;
+    case predictWinner.status.player:
+      playerResultElem.innerHTML = "Win!";
+      player.score++;
+      break;
+    case predictWinner.status.computer:
+      computerResultElem.innerHTML = "Win!";
+      computer.score++;
+      break;
+  }
+
+  setGamePoints();
+  endGame();
+}
+
+function playerPick(playerPick) {
+  var computerPick = getComputerPick();
+
+  playerPickElem.innerHTML = playerPick;
+  computerPickElem.innerHTML = computerPick;
+
+  checkRoundWinner(playerPick, computerPick);
+}
+
+function setGamePoints() {
+  playerPointsElem.innerHTML = player.score;
+  computerPointsElem.innerHTML = computer.score;
+}
+
+function endGame() {
+  if (player.score == 10) {
+    alert(player.name + ' WINS!');
+    gameState = 'ended';
+    setGameElements();
+  } else if (computer.score == 10) {
+    alert('Computer WINS!');
+    gameState = 'ended';
+    setGameElements();
+  }
+}
