+++ package.json
@@ -1,0 +1,26 @@
+{
+  "name": "9.7-Kodilla-Bootcamp",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "serve": "webpack-dev-server --progress --colors --open"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/stangricki/9.7-Kodilla-Bootcamp.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/stangricki/9.7-Kodilla-Bootcamp/issues"
+  },
+  "homepage": "https://github.com/stangricki/9.7-Kodilla-Bootcamp#readme",
+  "devDependencies": {
+    "html-webpack-plugin": "^2.30.1",
+    "webpack": "^3.8.1",
+    "webpack-dev-server": "^2.9.4"
+  },
+  "dependencies": {}
+}
