+++ webpack.config.js
@@ -1,0 +1,18 @@
+const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+  entry: {
+    app: './js/script.js',
+  },
+  output: {
+    filename: 'bundle.js',
+    path: path.resolve(__dirname, 'dist')
+  },
+  devtool: 'inline-source-map', 
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: 'index.html',
+    })
+  ],
+};
